Split the freezing vs. streaming tests of watching
